feat(student): 학생 전시 상세 페이지 내 TeamMemberSection API 연동 (#73)
