+++ src/pages/Exhibition/Student/Detail/TeamMemberSection/MemberCard/StudentMemberCard.styled.ts
